fix(tui): 修复长对话滚动溢出
多轮长对话后，消息区全局视觉行号和滚动 offset 使用 u16 保存，累计渲染行超过 65535 后会饱和，导致 follow-bottom 停在旧位置，新回复无法继续顶到最新底部，鼠标滚动也会被错误边界卡住。

修改内容：
- render_thread.rs 将 WrappedLineInfo 视觉行号改为 usize，避免 wrap_map 行号截断
- ui_state.rs/thread_ops.rs/agent_submit.rs 将消息区 scroll offset 与边界改为 usize，并保持 follow-bottom 语义
- message_area.rs 移除 u16 饱和滚动计算，并在渲染时消费 RebuildWithAnchor 产生的 scroll_anchor
- text_selection.rs/event/mod.rs 同步消息区文本选择坐标类型，避免长历史下选区行号截断
- main.rs/message_area_test.rs/render_thread_test.rs 补充大行号与滚动边界回归测试

特性/影响：
- 修复超长 TUI 对话中新回复不自动滚到底部、鼠标滚轮滚动边界异常的问题
- 面板滚动仍保持原局部 u16 逻辑，不扩大改动范围

+++ peri-tui/src/app/agent_render.rs
@@ -175,14 +175,14 @@ impl App {
 
                 let anchor_message_idx = {
                     let cache = self.session_mgr.current().messages.render_cache.read();
-                    let scroll_row = self.session_mgr.current().ui.scroll_offset as usize;
+                    let scroll_row = self.session_mgr.current().ui.scroll_offset;
                     let msg_idx = cache
                         .message_offsets
                         .iter()
                         .enumerate()
                         .find(|(_, &offset)| {
                             offset < cache.wrap_map.len()
-                                && cache.wrap_map[offset].visual_row_start as usize >= scroll_row
+                                && cache.wrap_map[offset].visual_row_start >= scroll_row
                         })
                         .map(|(idx, _)| idx)
                         .unwrap_or(prefix_len);
